requirements.txt
fastapi[all]
google-cloud-storage
python-multipart
Pillow
timm
opencv-python